BoardView display board
ConsoleDisplayView display into console string version of Board
