+++ src/main/java/com/maciejkocur/tictactoe/ui/BoardView.java
@@ -1,0 +1,8 @@
+package com.maciejkocur.tictactoe.ui;
+
+import com.maciejkocur.tictactoe.board.Board;
+
+public interface BoardView {
+
+    void display(Board board);
+}
