+++ src/main/java/com/maciejkocur/tictactoe/ui/impl/ConsoleBoardView.java
@@ -1,0 +1,12 @@
+package com.maciejkocur.tictactoe.ui.impl;
+
+import com.maciejkocur.tictactoe.board.Board;
+import com.maciejkocur.tictactoe.ui.BoardView;
+
+public class ConsoleBoardView implements BoardView {
+
+    @Override
+    public void display(Board board) {
+        System.out.println(board.toString());
+    }
+}
